+++ androidApp/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:label="@string/app_name"
         android:usesCleartextTraffic="true"
         android:networkSecurityConfig="@xml/network_security_config"
+        android:windowSoftInputMode="adjustResize"
         android:theme="@style/Theme.Yaba">
         <activity
             android:name=".MainActivity"
