+++ androidApp/src/main/java/tech/alexib/yaba/kmm/android/YabaBottomAppBar.kt
@@ -51,9 +51,9 @@ fun YabaBottomBar(navController: NavHostController) {
                 label = { Text(text = "Settings") },
                 onClick = {
                     navController.navigate(Route.Settings.route) {
-                        popUpTo(navController.graph.findStartDestination().id) {
-                            saveState = true
-                        }
+//                        popUpTo(navController.graph.findStartDestination().id) {
+//                            saveState = true
+//                        }
                         // Avoid multiple copies of the same destination when
                         // reselecting the same item
                         launchSingleTop = true
