--- androidApp/src/main/java/tech/alexib/yaba/kmm/android/ui/plaid_items/PlaidItemsViewModel.kt
@@ -1,20 +1,0 @@
-package tech.alexib.yaba.kmm.android.ui.plaid_items
-
-//class PlaidItemsViewModel : ViewModel(), KoinComponent {
-//
-//
-//    private val service: PlaidLinkService by inject()
-//
-//    fun linkInstitution(handleToken: (LinkTokenConfiguration) -> Unit) {
-//        viewModelScope.launch {
-//            service.linkInstitution(handleToken)
-//        }
-//    }
-//
-//
-//    fun handleResult(linkResult: LinkResult) {
-//        viewModelScope.launch {
-//            service.handleResult(linkResult)
-//        }
-//    }
-//}
